# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KashFlow API",
    "description": "Achat groupé à prix dégressif. Les paliers portent sur la quantité commandée, jamais sur le nombre de participants (D1), et le prix est rétroactif : quand un palier tombe, tout le groupe en profite (D2).",
    "version": "0.2.0"
  },
  "paths": {
...
    "/groups": {
      "get": {
        "tags": [
          "groupes"
        ],
        "summary": "List Groups",
        "operationId": "list_groups_groups_get",
        "parameters": [
          {
            "name": "product_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GroupCard"
                  },
                  "title": "Response List Groups Groups Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "groupes"
        ],
        "summary": "Create Group",
        "operationId": "create_group_groups_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupCreateIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/groups/{group_id}": {
      "get": {
        "tags": [
          "groupes"
        ],
        "summary": "Détail d'un groupe (endpoint du polling 2s)",
        "operationId": "get_group_groups__group_id__get",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Group Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupDetail"
                }
              }
            }
          },
          "404": {
            "description": "Groupe introuvable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/groups/code/{share_code}": {
      "get": {
        "tags": [
          "groupes"
        ],
        "summary": "Détail d'un groupe depuis un lien partagé (consultable sans compte)",
        "operationId": "get_group_by_code_groups_code__share_code__get",
        "parameters": [
          {
            "name": "share_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Share Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupDetail"
                }
              }
            }
          },
          "404": {
            "description": "Groupe introuvable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/groups/{group_id}/join": {
      "post": {
        "tags": [
          "groupes"
        ],
        "summary": "Join Group",
        "description": "Rejoindre : création de la commande, recalcul du palier et propagation\nrétroactive à tout le groupe, le tout dans une seule transaction.",
        "operationId": "join_group_groups__group_id__join_post",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Group Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinOut"
                }
              }
            }
          },
          "404": {
            "description": "Groupe introuvable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "Groupe fermé, déjà rejoint ou stock insuffisant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/groups/{group_id}/leave": {
      "post": {
        "tags": [
          "groupes"
        ],
        "summary": "Leave Group",
        "operationId": "leave_group_groups__group_id__leave_post",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Group Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupDetail"
                }
              }
            }
          },
          "404": {
            "description": "Groupe introuvable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorOut": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "code": {
            "type": "string",
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorOut",
        "description": "Réponse d'erreur uniforme : le front ne devine pas."
      },
      "GroupCard": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "share_code": {
            "type": "string",
            "title": "Share Code"
          },
          "status": {
            "$ref": "#/components/schemas/GroupStatus"
          },
          "participants_count": {
            "type": "integer",
            "title": "Participants Count"
          },
          "current_quantity": {
            "type": "integer",
            "title": "Current Quantity"
          },
          "target_quantity": {
            "type": "integer",
            "title": "Target Quantity"
          },
          "current_unit_price": {
            "type": "integer",
            "title": "Current Unit Price"
          },
          "progress_ratio": {
            "type": "number",
            "title": "Progress Ratio"
          },
          "seconds_remaining": {
            "type": "integer",
            "title": "Seconds Remaining"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "share_code",
          "status",
          "participants_count",
          "current_quantity",
          "target_quantity",
          "current_unit_price",
          "progress_ratio",
          "seconds_remaining"
        ],
        "title": "GroupCard"
      },
      "GroupCreateIn": {
        "properties": {
          "product_id": {
            "type": "integer",
            "title": "Product Id"
          },
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Name"
          },
          "target_quantity": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Target Quantity"
          },
          "min_quantity": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Min Quantity"
          },
          "deadline_hours": {
            "type": "integer",
            "maximum": 720.0,
            "minimum": 1.0,
            "title": "Deadline Hours",
            "default": 48
          },
          "quantity": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Quantity",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "name",
          "target_quantity",
          "min_quantity"
        ],
        "title": "GroupCreateIn"
      },
      "GroupDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "share_code": {
            "type": "string",
            "title": "Share Code"
          },
          "status": {
            "$ref": "#/components/schemas/GroupStatus"
          },
          "deadline": {
            "type": "string",
            "format": "date-time",
            "title": "Deadline"
          },
          "seconds_remaining": {
            "type": "integer",
            "title": "Seconds Remaining"
          },
          "product": {
            "$ref": "#/components/schemas/GroupProductOut"
          },
          "participants_count": {
            "type": "integer",
            "title": "Participants Count"
          },
          "current_quantity": {
            "type": "integer",
            "title": "Current Quantity"
          },
          "target_quantity": {
            "type": "integer",
            "title": "Target Quantity"
          },
          "min_quantity": {
            "type": "integer",
            "title": "Min Quantity"
          },
          "current_unit_price": {
            "type": "integer",
            "title": "Current Unit Price"
          },
          "current_tier": {
            "$ref": "#/components/schemas/TierOut"
          },
          "next_tier": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TierOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "quantity_to_next_tier": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity To Next Tier"
          },
          "progress_ratio": {
            "type": "number",
            "title": "Progress Ratio"
          },
          "unit_saving": {
            "type": "integer",
            "title": "Unit Saving"
          },
          "potential_unit_saving": {
            "type": "integer",
            "title": "Potential Unit Saving"
          },
          "group_total_saving": {
            "type": "integer",
            "title": "Group Total Saving"
          },
          "my_membership": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MembershipOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "share_code",
          "status",
          "deadline",
          "seconds_remaining",
          "product",
          "participants_count",
          "current_quantity",
          "target_quantity",
          "min_quantity",
          "current_unit_price",
          "current_tier",
          "next_tier",
          "quantity_to_next_tier",
          "progress_ratio",
          "unit_saving",
          "potential_unit_saving",
          "group_total_saving",
          "my_membership"
        ],
        "title": "GroupDetail"
      },
      "GroupProductOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "unit_label": {
            "type": "string",
            "title": "Unit Label"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "individual_price": {
            "type": "integer",
            "title": "Individual Price"
          },
          "merchant_name": {
            "type": "string",
            "title": "Merchant Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "unit_label",
          "image_url",
          "individual_price",
          "merchant_name"
        ],
        "title": "GroupProductOut"
      },
      "GroupStatus": {
        "type": "string",
        "enum": [
          "OPEN",
          "LOCKED",
          "COMPLETED",
          "CANCELLED"
        ],
        "title": "GroupStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JoinIn": {
        "properties": {
          "quantity": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Quantity"
          }
        },
        "type": "object",
        "required": [
          "quantity"
        ],
        "title": "JoinIn"
      },
      "JoinOut": {
        "properties": {
          "order": {
            "$ref": "#/components/schemas/OrderOut"
          },
          "group": {
            "$ref": "#/components/schemas/GroupDetail"
          },
          "tier_unlocked": {
            "type": "boolean",
            "title": "Tier Unlocked"
          },
          "previous_unit_price": {
            "type": "integer",
            "title": "Previous Unit Price"
          }
        },
        "type": "object",
        "required": [
          "order",
          "group",
          "tier_unlocked",
          "previous_unit_price"
        ],
        "title": "JoinOut"
      },
...
      "MembershipOut": {
        "properties": {
          "joined": {
            "type": "boolean",
            "title": "Joined"
          },
          "order_id": {
            "type": "integer",
            "title": "Order Id"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "total_amount": {
            "type": "integer",
            "title": "Total Amount"
          }
        },
        "type": "object",
        "required": [
          "joined",
          "order_id",
          "quantity",
          "total_amount"
        ],
        "title": "MembershipOut"
      },
...
      "OrderOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "group_id": {
            "type": "integer",
            "title": "Group Id"
          },
          "group_name": {
            "type": "string",
            "title": "Group Name"
          },
          "product_id": {
            "type": "integer",
            "title": "Product Id"
          },
          "product_name": {
            "type": "string",
            "title": "Product Name"
          },
          "unit_label": {
            "type": "string",
            "title": "Unit Label"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "unit_price": {
            "type": "integer",
            "title": "Unit Price"
          },
          "total_amount": {
            "type": "integer",
            "title": "Total Amount"
          },
          "individual_price": {
            "type": "integer",
            "title": "Individual Price"
          },
          "saving": {
            "type": "integer",
            "title": "Saving"
          },
          "payment_status": {
            "$ref": "#/components/schemas/PaymentStatus"
          },
          "order_status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "group_id",
          "group_name",
          "product_id",
          "product_name",
          "unit_label",
          "quantity",
          "unit_price",
          "total_amount",
          "individual_price",
          "saving",
          "payment_status",
          "order_status",
          "created_at"
        ],
        "title": "OrderOut"
      },
      "OrderStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "CONFIRMED",
          "CANCELLED"
        ],
        "title": "OrderStatus"
      },
...
      "PaymentStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "SUCCESS",
          "FAILED",
          "REFUNDED"
        ],
        "title": "PaymentStatus"
      },
...
      "TierOut": {
        "properties": {
          "min_quantity": {
            "type": "integer",
            "title": "Min Quantity"
          },
          "unit_price": {
            "type": "integer",
            "title": "Unit Price"
          }
        },
        "type": "object",
        "required": [
          "min_quantity",
          "unit_price"
        ],
        "title": "TierOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}